+++ src/blocks/HamburgerMenu/index.ts
@@ -1,3 +1,0 @@
-export { HamburgerMenu } from "./HamburgerMenu";
-export type { HamburgerMenuProps } from "./HamburgerMenu";
-