+++ src/components/GlobalNavMenu/index.ts
@@ -1,8 +1,0 @@
-export { GlobalNavMenu, productBlockVariants } from "./GlobalNavMenu";
-export type {
-  GlobalNavMenuProps,
-  NavMenuItem,
-  ProductBlockProps,
-  DesktopMenuContentProps,
-  MobileMenuContentProps,
-} from "./GlobalNavMenu";